backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
celery==5.3.4
redis==5.0.1
httpx==0.25.1
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cloudinary==1.36.0
boto3==1.29.7
faiss-cpu==1.7.4
annoy==1.17.3
